- changed setupNetwork* functions in a way that they check   for the presence of the $root device and in case of an   error and an existing root device LOCAL_BOOT is set and   the system try to proceed with local boot of the installed   system (fate #313338)

+++ modules/KIWILinuxRC.sh
@@ -3576,9 +3576,16 @@ function setupNetwork {
 		unset opts
 	fi
 	if ! dhcpcd $opts $PXE_IFACE 1>&2;then
-		systemException \
-			"Failed to setup DHCP network interface !" \
-		"reboot"
+		if [ -e "$root" ];then
+			Echo "Failed to setup DHCP network interface !"
+			Echo "Try fallback to local boot on: $root"
+			LOCAL_BOOT=yes
+			return
+		else
+			systemException \
+				"Failed to setup DHCP network interface !" \
+			"reboot"
+		fi
 	fi
 	for i in 1 2 3 4 5 6 7 8 9 10 11 12 13 14 15 16 17 18 19 20;do
 		if [ -s /var/lib/dhcpcd/dhcpcd-$PXE_IFACE.info ];then
@@ -3587,9 +3594,16 @@ function setupNetwork {
 		sleep 2
 	done
 	if [ ! -s /var/lib/dhcpcd/dhcpcd-$PXE_IFACE.info ];then
-		systemException \
-			"Can't find DHCP info file: dhcpcd-$PXE_IFACE.info !" \
-		"reboot"
+		if [ -e "$root" ];then
+			Echo "Can't find DHCP info file: dhcpcd-$PXE_IFACE.info !"
+			Echo "Try fallback to local boot on: $root"
+			LOCAL_BOOT=yes
+			return
+		else
+			systemException \
+				"Can't find DHCP info file: dhcpcd-$PXE_IFACE.info !" \
+			"reboot"
+		fi
 	fi
 	ifconfig lo 127.0.0.1 netmask 255.0.0.0 up
 	for i in 1 2 3 4 5 6 7 8 9 10 11 12 13 14 15 16 17 18 19 20;do
@@ -3600,9 +3614,16 @@ function setupNetwork {
 		sleep 2
 	done
 	if [ -z "$IPADDR" ];then
-		systemException \
-			"Can't assign IP address from dhcp info: dhcpcd-$PXE_IFACE.info !" \
-		"reboot"
+		if [ -e "$root" ];then
+			Echo "Can't assign IP addr via dhcp info: dhcpcd-$PXE_IFACE.info !"
+			Echo "Try fallback to local boot on: $root"
+			LOCAL_BOOT=yes
+			return
+		else
+			systemException \
+				"Can't assign IP addr via dhcp info: dhcpcd-$PXE_IFACE.info !" \
+				"reboot"
+		fi
 	fi
 	if [ -z "$DOMAIN" ] && [ -n "$DNSDOMAIN" ];then
 		export DOMAIN=$DNSDOMAIN
